scripts/python/requirements.txt
pymongo
playwright
python-dotenv
requests
pandas
matplotlib
openpyxl
